improved submodule cohesion
