+++ src/adverts/adverts.gql.schema.ts
