+++ src/adverts/e2e-test/create-advert.spec.ts
@@ -1,7 +1,7 @@
-import { end2endTest } from '../test-utils'
+import { end2endTest } from '../../test-utils'
 import { StatusCodes } from 'http-status-codes'
 import { createAdvertMutation } from './queries'
-import { AdvertInput } from '../adverts/types'
+import { AdvertInput } from '../types'
 
 describe('createAdvert', () => {
 	it('xxx', () => end2endTest(async ({ gqlRequest }) => {
