+++ src/adverts/e2e-test/queries.ts
