+++ src/terms/terms.gql.schema.ts
